move theme component
